Extract the workstation trading payload into a partial
Increment 2m of the endpoint-monolith decomposition. Moves the trading workspace
payload builders (BuildTradingPayloadAsync, and the fixture BuildTradingFallbackPayload
/ BuildTradingFallbackReadiness) out of
src/Meridian.Ui.Shared/Endpoints/WorkstationEndpoints.cs into a new sibling
partial WorkstationEndpoints.Trading.cs.

Behavior-preserving verbatim relocation: closed call graph (the three methods
have zero references outside WorkstationEndpoints.cs; the only entry is the
inline GetWorkstationTrading route lambda in core). The shared helpers in the
gap and elsewhere — GetTradingOperatorReadinessAsync, NormalizeOperatorInboxToken,
BuildTradingBrokerageNotes, ResolveLiveMark, ResolveRuntimeRiskDescriptorAsync,
BuildModeComparisons, BuildRunDrillInLinks, FormatCurrency/FormatPercent — remain
in core and are reached across the partial. Full using-audit performed against
each referenced type's namespace.

Core WorkstationEndpoints.cs drops 5,384 -> 5,084 lines; ratchet baseline
retightened.
